feat: add AddRemoveRelationsIdempotent (#58)
* feat(relation): add idempotent methods to add and remove relations

Enable the use of ConflicOption release in Openfga v1.10.0 to handle idempotent add/remove. I've
used an option pattern because it seems aligned with the way the go-sdk adds new features, so it
will help us in the future when they'll add new features.

* test(integration-test): add integration tests against a real openfga instance

add integration tests against an openfga instance, with instructions on how to run them locally